+++ go.mod
@@ -2,4 +2,13 @@ module github.com/minuk-dev/otelcol-config-lint
 
 go 1.25.4
 
-require gopkg.in/yaml.v3 v3.0.1
+require (
+	github.com/spf13/cobra v1.10.2
+	gopkg.in/yaml.v3 v3.0.1
+)
+
+require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/samber/mo v1.17.0 // indirect
+	github.com/spf13/pflag v1.0.9 // indirect
+)
