--- internal/cli/columns.go
@@ -1,30 +1,0 @@
-package cli
-
-import (
-	"io"
-	"text/tabwriter"
-)
-
-// columns renders aligned help output for -list-rules and -list-versions.
-type columns struct{ w *tabwriter.Writer }
-
-// columnPadding is the gap between columns, in spaces.
-const columnPadding = 2
-
-func newColumns(w io.Writer) columns {
-	return columns{w: tabwriter.NewWriter(w, 0, 0, columnPadding, ' ', 0)}
-}
-
-func (c columns) row(cells ...string) {
-	for i, cell := range cells {
-		if i > 0 {
-			_, _ = io.WriteString(c.w, "\t")
-		}
-
-		_, _ = io.WriteString(c.w, cell)
-	}
-
-	_, _ = io.WriteString(c.w, "\n")
-}
-
-func (c columns) flush() { _ = c.w.Flush() }
